+++ hypervisor/Makefile
@@ -151,6 +151,7 @@ endif
 COMMON_C_SRCS += common/notify.c
 COMMON_C_SRCS += lib/memory.c
 COMMON_C_SRCS += common/percpu.c
+COMMON_C_SRCS += common/cpu.c
 
 ifeq ($(ARCH),x86)
 COMMON_C_SRCS += common/ticks.c
