+++ hypervisor/arch/x86/lapic.c
@@ -14,6 +14,7 @@
 #include <asm/apicreg.h>
 #include <asm/irq.h>
 #include <delay.h>
+#include <cpu.h>
 
 /* intr_lapic_icr_delivery_mode */
 #define INTR_LAPIC_ICR_FIXED           0x0U
