+++ hypervisor/debug/sbuf.c
@@ -14,6 +14,7 @@
 #include <errno.h>
 #include <asm/cpu.h>
 #include <per_cpu.h>
+#include <cpu.h>
 
 int32_t sbuf_share_setup(uint16_t pcpu_id, uint32_t sbuf_id, uint64_t *hva)
 {
